+++ packages/evaluator/src/cspEvaluators.ts
@@ -1,0 +1,4 @@
+import { createEvaluators } from './createEvaluators';
+import { cspFunctionFromExpression } from './cspFunctionFromExpression';
+
+export const cspEvaluate = createEvaluators(cspFunctionFromExpression);
