+++ packages/evaluator/src/index.ts
@@ -1,3 +1,5 @@
-export { evaluate } from './evaluate';
-export { evaluateLater } from './evaluateLater';
+export { createEvaluators } from './createEvaluators';
+export { cspEvaluate } from './cspEvaluators';
+export { cspFunctionFromExpression } from './cspFunctionFromExpression';
 export { functionFromExpression } from './functionFromExpression';
+export { evaluate, evaluateLater } from './unsafeEvaluators';
