--- packages/evaluator/src/nextTick.ts
@@ -1,2 +1,0 @@
-export const nextTick = () =>
-  new Promise<void>((res) => queueMicrotask(() => setTimeout(res, 0)));
