+++ packages/evaluator/src/unsafeEvaluators.ts
@@ -1,0 +1,6 @@
+import { createEvaluators } from './createEvaluators';
+import { functionFromExpression } from './functionFromExpression';
+
+const { evaluate, evaluateLater } = createEvaluators(functionFromExpression);
+
+export { evaluate, evaluateLater };
